Feature: Created the patients app'
