+++ src/mcare_backend/authapp/models.py
@@ -81,17 +81,6 @@ class CustomUser(AbstractBaseUser):
         return True
 
 
-class PatientProfile(models.Model):
-    user = models.OneToOneField(settings.AUTH_USER_MODEL,
-                                limit_choices_to={'role': 'Patient'},
-                                on_delete=models.CASCADE,
-                                related_name="patient_profile")
-
-    # your custom fields for Patient model
-    def __str__(self):
-        return self.user.username
-
-
 class ExpertProfile(models.Model):
     user = models.OneToOneField(settings.AUTH_USER_MODEL,
                                 limit_choices_to={'role': 'Expert'},
