+++ src/mcare_backend/mcare_backend/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = [
 
     # installed apps
     'authapp',
+    'patients',
 
     # for token authentication
     'rest_framework.authtoken',
@@ -56,7 +57,7 @@ REST_FRAMEWORK = {
         'rest_framework.authentication.TokenAuthentication',
     ],
     'DEFAULT_PERMISSION_CLASSES': [
-        'rest_framework.permissions.IsAuthenticated',
+        # 'rest_framework.permissions.IsAuthenticated',
     ]
 }
