+++ src/mcare_backend/mcare_backend/urls.py
@@ -16,7 +16,12 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 
+
+from patients.urls import router
+
+
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/v1/authuser/', include('authapp.urls'))
+    path('api/v1/authuser/', include('authapp.urls')),
+    path('api/v1/', include(router.urls)),
 ]
