+++ src/mcare_backend/patients/admin.py
@@ -1,0 +1,13 @@
+from django.contrib import admin
+from .models import PatientProfile
+
+# Register your models here.
+
+
+class PatientProfileAdmin(admin.ModelAdmin):
+    '''Register the custom user model'''
+
+    model = PatientProfile
+
+
+admin.site.register(PatientProfile, PatientProfileAdmin)
