+++ src/mcare_backend/patients/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PatientsConfig(AppConfig):
+    name = 'patients'
