+++ src/mcare_backend/patients/serializers.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+from patients.models import PatientProfile
+
+
+class PatientProfileSerializer(serializers.ModelSerializer):
+
+    owner = serializers.HiddenField(
+        default=serializers.CurrentUserDefault()
+    )
+
+    class Meta:
+        model = PatientProfile
+        fields = ('id', 'owner')
